moves styled components directly to filter
